requirements.txt
torch>=2.0.1
torchvision>=0.15.2
faster-coco-eval>=1.6.5
PyYAML
tensorboard
scipy
calflops
transformers
defusedxml
supervision
onnxruntime-gpu
onnx 
loguru